bnp: work in progress: new db source (turtle)
There's another database with BNP's data, this one with all the info
we want.

Using this source, we can dump the other two. This scripts aren't
finished yet (process-bnp.sh needs to be completed), but when they
are, we can move this from bnp/turtle to bnp, and change the README
and the quem-morreu script in order to properly use the new processed
files.

Ref.: #7