csv2wiki: calculate PD year from YoD
The CSV is supposed to have a list of authors, all dead in the same year, all
entering Public Domain 70 years afterwards. Instead of having the year of entry
into Public Domain hardcoded in the script, it is now calculated by reading the
death year of one of the entries in the CSV file.