wikidata query: add range-safe predicate to date
We're now informing the query service that date of death is a
range-safe predicate. This tells the optimizer that it doesn’t mix
dates, strings, integers or other data types, which simplifies the
range comparison. This is almost always true on Wikidata, so the main
reason not to use this optimizer hint all the time is that it’s a bit
inconvenient. (It’s probably also not a good idea to use it when
you’re working with unknown values.)

More info:
https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service/query_optimization#Fixed_values_and_ranges